# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTripConfigRequest:
    properties:
      arrival_time:
        type: string
      departure_time:
        type: string
      duration_days:
        type: integer
      expiration_date:
        type: string
      recurrence:
        type: string
      route_id:
        type: string
      ship_id:
        type: string
      start_date:
        type: string
    required:
    - route_id
    - ship_id
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: detailed error message
        type: string
    type: object
...
  dto.SuccessMessageResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.TripConfigResponse:
    properties:
      created_at:
        type: string
      departure_time:
        type: string
      duration_days:
        type: integer
      expiration_date:
        type: string
      id:
        type: string
      recurrence:
        type: string
      route_id:
        type: string
      route_name:
        type: string
      ship_id:
        type: string
      ship_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Service for managing user identities, roles, and organizations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Uber dos Rios API
  version: v1.0.16
paths:
...
  /trips-config:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of trips.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TripConfigResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List Trip Config
      tags:
      - trips-config
    post:
      consumes:
      - application/json
      description: Registers a new Trip Config with the provided details.
      parameters:
      - description: Trip Config Data to Create
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTripConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TripConfigResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new Trip Config
      tags:
      - trips-config
  /trips-config/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a Trip Config by its ID.
      parameters:
      - description: Trip Config ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessMessageResponse'
        "400":
          description: Invalid Trip Config ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a Trip Config
      tags:
      - trips-config
...
schemes:
- http
- https
swagger: "2.0"